refactor constants + other changes
